add scopes.check_scope_filter
Extracted from APIHandler.get_scope_filter for easier re-use

and mve get_scope_filter to BaseHandler from APIHandler since it will be needed on oauth
